+++ src/transformers/models/gpt_megatron/activations/glu.py
