--- src/transformers/models/gpt_megatron/modeling_utils/__init__.py
@@ -1,13 +1,0 @@
-from .activations import get_activation_function, is_glu
-from .attention import (
-    SDPA,
-    Attention,
-    FlashAttention,
-    MathAttention,
-    PackedFlashAttention,
-    get_attention_module,
-    interleave_query_key_value_tensor_for_attention,
-    split_query_key_value_tensor_for_attention,
-)
-from .normalization import RMSNorm, get_normalization_function
-from .position_embedding import Alibi, RoPE, YaRNScaledRoPE
