+++ src/transformers/models/gpt_megatron/normalization/rmsnorm/apex.py
