+++ src/transformers/models/gpt_megatron/normalization/rmsnorm/base.py
